AB#1254 Move modal out of show page into approvals
